requirements.txt
discord.py>=2.3
httpx>=0.27
beautifulsoup4>=4.12
